requirements_openai.txt
SpeechRecognition>=3.8.1
pyttsx3==2.90
PyAudio
pocketsphinx
openai>=1.0.0
pyobjc; platform_system == "Darwin"
python-dotenv